scope-osc-prompt/pyproject.toml
[project]
name = "scope-osc-prompt"
version = "0.1.0"
description = "Receive OSC messages and inject them as diffusion prompts in Daydream Scope"
requires-python = ">=3.12"
dependencies = ["python-osc", "scope-bus"]

[project.entry-points."scope"]
scope_osc_prompt = "scope_osc_prompt"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scope_osc_prompt"]
